Separate out components from main file
