+++ src/components/ProjectShowcase/components/Banner/index.jsx
@@ -1,0 +1,15 @@
+import * as React from 'react';
+
+function Banner() {
+  return (
+    <div className="project-portal__banner">
+      <h1>Chingu</h1>
+      <p>
+        Chingu facilitates global collaboration on projects in a structure
+        focused on improving remote development skills
+      </p>
+    </div>
+  );
+}
+
+export default Banner;
